Move CgroupV1KubernetesContainerUtil into hono-client-command.
